Extract name and Evalue of top hit
My plan is to extract Evalues for the top blast hit of each Chloroplast
development gene,
then blast those sequences against nr to and note the function of the
top hit.

So far I've done all of the blasting of candidate genes and extracted
the sequences.

This commit extracts the evalues.

Next I need a way to run a nr blast search that will produce top hits
with informative
names (not something like Hypothetical protein)