fix strand flips and filter ambiguous SNPs
Turns out the the output of checkVCF.py mixes strand flips and allele
flips.

I had to hack the code of the script to differentiate these, and while
I was at it, I modified the code to include the SNP ID in the output
along with the genomic coordinate. That makes all of the joining that I
spent the morning figuring out unnecessary.