final output copied to separate file
This was necessary because there are WAY too many files to pick out the
ones to upload to the imputation server. For some reason, the final
gzip files were fubar. I think I ran these on my laptop and something
clearly went wrong. When I reran it on my iMac everything is fine