added code to shorten names and make FastQC heat map bigger
I worked on a perl regex to further shorten these names, but it turns
out that the whole thing is needed to make the unique. It looks like
the sample numbers are not encoded in the file names but only in the
enclosing folders. Since that info is no longer with the files, I will
have to go back to the original raw data to retrieve it