Dealing with SNP IDs
- Output from the Michigan Imputation server does not include the rsID
of the SNPs.
- Nick would like to keep this information in the VCF files. I think it
can’t hurt since soaring the chromosome and position in the ID field is
pretty redundant
- It is easy to get SNP positions and rsIDs from dbSNP, but there can
be multiple SNPs at the same position
- This can be solved by only taking the one where the alleles match
what is in out VCFs (and taking the lower rsID if there are true
duplicates
- Allele information is not easy to extract from SNPdb however.
- I have managed to download it from UCSC (along with a bonus DB
schema), but it is taking a while to import the 26 GB file into mySQL
(I will have to be careful to erase the original once it finishes)
- once that’s done, it should be easy to match up the SNPs from the
VCF. Whether or not it’s worth actually updating the VCFs I have not
yet decided.
- At present, the AddSNPID.py script just compares the rsIDs from the
VCF to the ones in the DB. I need to modify it to actually write a new
VCF. I also have to change it to use the UCSC data and to include the
ref and Alt allele in the lookup. Note that there may be allele swaps
that have to be dealt with when I do this