Redid subsetting and concatating of imputed VCF files
- Turns out I had used the wrong identifier for one of the samples for
the data that I sent to Nick’s collaborator. Good thing I used a script
to change the names rather than doing it manually because it means the
VCF wound up with the wrong ID as well as the wrong data rather than
just the wrong data!

- I had to redo the concatenation because I wound up with the
chromosomes in the wrong order when I concatenated everything.