modified to select rsID based on genotype and position
This seems to work EXCEPT that the omni chip data has a lot of
undercalls where the alt allele is listed as ‘.’, but the alt genotype
for that SNP is a base.

I’ve also modified it to search based on the start position of the
variant rather than the end position because variants that span
multiple positions in the ref aren’t correct (VCF uses 1-based start
position, DB uses 0-based start and 1-based end). I haven’t been able
to test this because I need to index the chromosome start position in
the DB