Change from hg19 to GRCh38Decoy
As Nick pointed out, according to Weirick et al., more reads map to the
newer assembly. It is also apparently important to include as much
sequence as possible to prevent reads from parts of the genome that
aren’t part of the assembly incorrectly mapping. This assembly has
decoy sequences in it to do this.

It’s also going to be important to ensure that the features
corresponding to this assembly are used in the downstream QC steps.
This probably means converting the GTF file to BED format for RSeQC.
BAMqc and Picard tools require other tweaks, but I think this RSeQC is
the most useful of the bunch