Figure out why pairs aren't mapping
According to both RSeQC and Picard tools, our data are second strand,
but RSeQC reports very few reads properly paired when I use the
fr-secondstrand option in TopHat. BAMqc reports that most of the reads
are properly paired, but if I run Samtools pileup without the -A flag,
I get very few bases. When I use fr-firststrand or fr-unstranded, I
also get very few properly paired reads (and very little pileup without
the -A flag). Judging by the insert size distribution, this might be
because I used to large a size, but that doesn’t really make sense
because it looks like only validly paired reads were used for that
distribution. Also, Eilis used an even larger insert size, and I get
plenty of bases in the pileup when I run it on her data without the -A
flag. I’m trying it with a smaller insert size and we’ll see if that
makes a difference