Modified to add IDs rather than checking them against VCF
IDs are identified by position (start and start plus ref length), if
both ref and alt from the VCF (or there reverse-complements) are
present among the observed variants. If there are multiple matching SNP
IDs the first one in the DB is used. If the SNPs are flagged as
duplicates, this is done silently, but a warning is printed if they are
not listed as duplicates in the DB.

In testing earlier version against the omni chip VCF, this worked in
all cases except a few where the omnichip SNP ID has been discontinued
and replaced by the DB one.