Modified WASP to work on Non-ref position only
After much consideration, I’ve decided that it doesn’t make sense to
include SNPs that aren’t present in a given sample (because SNPs that
aren’t present can’t affect mapping), but it does make sense to include
homozygous non-ref SNPs because if that are present in the same read as
a heterozygous SNP, they can bias mapping (admittedly, this is probably
a rare situation because LD). It may be that considering all SNPs makes
more sense when no considering allele-specific expression, because a
variant present in some samples may cause reduced mapping efficiency,
which will look like DE, so it’s probably worth doing it both ways.