Call SNPs from chr22 mappings and compare to VCF
This produces a plot for each sample showing which sample from the
omnichip VCF the RNAseq data most closely match. If it’s not the same
sample, there’s been a mixup.

I’ve modified these scripts to recapitulate the analysis on all
samples, which is not the way I’ve actually done them. I may redo these
steps to make sure that what I wrote produces the output it should.

I’ve also decided to rename sample 15533_2 and all the new ones that I
got this week. This is going to be a pain, but it’s going to save me
headaches in the long run. By using ‘-2’ for all duplicated samples,
rather than ‘_2’, I can strip off everything after the first
underscore, which will make it easy to get rid of the ‘_S##’ notation
that is included in the file names from the sequencer