add mapping QC and read counting steps
This just uses bash commands to add additional submission scripts to
the same analysis.

Some of these could be run in parallel, but they don’t take massive
amounts of time, and I’m usually getting new datasets in large
multiples, so I’m already using a lot of computation resources
simultaneously.

I would like to add the WASP analysis to this, but I’m currently
looking up the VCF index manually. It should be possible to automate
this (using grep maybe?), but I have not done so yet.

An additional step that I will need to add is automatically identifying
duplicates and changing the sampleID when they are identified