refactor using pysam, and filter on alt allele probabilities
This should make for hopefully faster and certainly faster VCF parsing.

I haven’t managed to figure out how to modify pysam objects, so it’s
necessary to just print everything, then compress with bcftools view

bcftools always complains because the chromosome ID is not specified in
the header. I’ve added that info (currently hard coded as ’22’), which
entails printing each line of the header rather than using the
print(bcf_in.header)

I had to use the print function from python3 because the python2.7
print function adds a new line after each print statement (or a space
if there’s a comma after the print statement). The VCF records already
have newlines, so needed to be able to specify “end=‘’” in the print
statement.