change to GRCh38alt, add sorting and verification steps
I suspect that the liftover outputs variant positions in the order that
they occur in the (sorted) input. Since the order can change after
conversion to the new coordinates, they need to be resorted.

I am using a chain file from UCSC, so I was getting errors from using
the decoy sequence. I’m now using the USCS version of the sequence

I added a checkVCF step to this script. It will be necessary to filter
out any positions that do not match the reference manually.