shorten project name for report
For some reason writeReport.edgeR is failing with a message of no
project name. Works fine when I analyse a single week. I’m assuming
that this is because the name is too long? Since it’s nested in a
folder with the full name, I decided to shorten the report name