Submit each step as a separate job, rather than running as bash commands
This allows different resources to be called for each step rather than
requiring 8 nodes for all steps including ones that are not parallelised.

This makes use of named jobs and the -hold_jid qsub argument.

I've commented out all of the logging of each step because thes qsub
commands complete immediately rather than waiting for the step to complete
as they did when they were bash commands.

I've also added comments to indicate the CPU and Wallclock times for each
step becasuse there appear to be some that are using more resources than
I'm asking for