make ASE counting separate step from combining mappings
This was initially split out as a way to avoid running CombineMappings
when there is only one mapping, but I ultimately decided to add code to
CombineMappings that automatically deals with renaming files in cases
where there is only a single mapping.

I decided to leave ASE counting as a separate step though, because it
allows me to do the counting for each chromosome in parallel

I also added code to CombineMappings to delete unnecessary intermediate
files, but I decided to comment these lines out so that they can be run
manually, making sure that all steps of the pipeline have run correctly
first.